Move classes into `SettingsTextBox`
