--- osu.Game/Graphics/UserInterface/OsuModSettingsNumberBox.cs
@@ -1,10 +1,0 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-namespace osu.Game.Graphics.UserInterface
-{
-    public class OsuModSettingsNumberBox : OsuModSettingsTextBox
-    {
-        protected override bool CanAddCharacter(char character) => char.IsNumber(character);
-    }
-}
